+++ eggs/Naaya/trunk/naaya/core/backport.py
@@ -1,15 +1,18 @@
 # Library code from newer versions of Python
 
+# do not delete, used for imports in several places
 try:
     import simplejson as json
 except ImportError:
     import json
+
 
 def all(l):
         for i in l:
             if not i:
                 return False
         return True
+
 
 def any(iterable):
     for element in iterable:
@@ -29,59 +32,72 @@ except ImportError:
         """Returns a new subclass of tuple with named fields.
 
         >>> Point = namedtuple('Point', 'x y')
-        >>> Point.__doc__                   # docstring for the new class
+        >>> Point.__doc__        # docstring for the new class
         'Point(x, y)'
-        >>> p = Point(11, y=22)             # instantiate with positional args or keywords
-        >>> p[0] + p[1]                     # indexable like a plain tuple
+        >>> p = Point(11, y=22)  # instantiate with positional args or keywords
+        >>> p[0] + p[1]          # indexable like a plain tuple
         33
-        >>> x, y = p                        # unpack like a regular tuple
+        >>> x, y = p             # unpack like a regular tuple
         >>> x, y
         (11, 22)
-        >>> p.x + p.y                       # fields also accessable by name
+        >>> p.x + p.y            # fields also accessable by name
         33
-        >>> d = p._asdict()                 # convert to a dictionary
+        >>> d = p._asdict()      # convert to a dictionary
         >>> d['x']
         11
-        >>> Point(**d)                      # convert from a dictionary
+        >>> Point(**d)           # convert from a dictionary
         Point(x=11, y=22)
-        >>> p._replace(x=100)               # _replace() is like str.replace() but targets named fields
+        >>> p._replace(x=100)    # _replace() is like str.replace()
+                                 # but targets named fields
         Point(x=100, y=22)
 
         """
 
         # Parse and validate the field names.  Validation serves two purposes,
-        # generating informative error messages and preventing template injection attacks.
+        # generating informative error messages and preventing template
+        # injection attacks.
         if isinstance(field_names, basestring):
-            field_names = field_names.replace(',', ' ').split() # names separated by whitespace and/or commas
+            # names separated by whitespace and/or commas
+            field_names = field_names.replace(',', ' ').split()
         field_names = tuple(map(str, field_names))
         if rename:
             names = list(field_names)
             seen = set()
             for i, name in enumerate(names):
-                if (not min(c.isalnum() or c=='_' for c in name) or _iskeyword(name)
-                    or not name or name[0].isdigit() or name.startswith('_')
-                    or name in seen):
-                        names[i] = '_%d' % i
+                if (not min(c.isalnum() or c == '_' for c in name)
+                        or _iskeyword(name)
+                        or not name or name[0].isdigit()
+                        or name.startswith('_')
+                        or name in seen):
+                    names[i] = '_%d' % i
                 seen.add(name)
             field_names = tuple(names)
         for name in (typename,) + field_names:
-            if not min(c.isalnum() or c=='_' for c in name):
-                raise ValueError('Type names and field names can only contain alphanumeric characters and underscores: %r' % name)
+            if not min(c.isalnum() or c == '_' for c in name):
+                raise ValueError(
+                    'Type names and field names can only contain '
+                    'alphanumeric characters and underscores: %r' % name)
             if _iskeyword(name):
-                raise ValueError('Type names and field names cannot be a keyword: %r' % name)
+                raise ValueError(
+                    'Type names and field names cannot be a keyword: '
+                    '%r' % name)
             if name[0].isdigit():
-                raise ValueError('Type names and field names cannot start with a number: %r' % name)
+                raise ValueError(
+                    'Type names and field names cannot start with a number: '
+                    '%r' % name)
         seen_names = set()
         for name in field_names:
             if name.startswith('_') and not rename:
-                raise ValueError('Field names cannot start with an underscore: %r' % name)
+                raise ValueError(
+                    'Field names cannot start with an underscore: %r' % name)
             if name in seen_names:
                 raise ValueError('Encountered duplicate field name: %r' % name)
             seen_names.add(name)
 
         # Create and fill-in the class template
         numfields = len(field_names)
-        argtxt = repr(field_names).replace("'", "")[1:-1]   # tuple repr without parens or quotes
+        # tuple repr without parens or quotes
+        argtxt = repr(field_names).replace("'", "")[1:-1]
         reprtxt = ', '.join('%s=%%r' % name for name in field_names)
         template = '''class %(typename)s(tuple):
         '%(typename)s(%(argtxt)s)' \n
@@ -94,7 +110,8 @@ except ImportError:
             'Make a new %(typename)s object from a sequence or iterable'
             result = new(cls, iterable)
             if len(result) != %(numfields)d:
-                raise TypeError('Expected %(numfields)d arguments, got %%d' %% len(result))
+                raise TypeError(
+                    'Expected %(numfields)d arguments, got %%d' %% len(result))
             return result \n
         def __repr__(self):
             return '%(typename)s(%(reprtxt)s)' %% self \n
@@ -102,10 +119,11 @@ except ImportError:
             'Return a new dict which maps field names to their values'
             return dict(zip(self._fields, self)) \n
         def _replace(_self, **kwds):
-            'Return a new %(typename)s object replacing specified fields with new values'
+            'Return an %(typename)s object replacing some fields'
             result = _self._make(map(kwds.pop, %(field_names)r, _self))
             if kwds:
-                raise ValueError('Got unexpected field names: %%r' %% kwds.keys())
+                raise ValueError(
+                    'Got unexpected field names: %%r' %% kwds.keys())
             return result \n
         def __getnewargs__(self):
             return tuple(self) \n\n''' % locals()
@@ -115,7 +133,8 @@ except ImportError:
             print template
 
         # Execute the template string in a temporary namespace
-        namespace = dict(_itemgetter=_itemgetter, __name__='namedtuple_%s' % typename,
+        namespace = dict(_itemgetter=_itemgetter,
+                         __name__='namedtuple_%s' % typename,
                          _property=property, _tuple=tuple)
         try:
             exec template in namespace
@@ -123,49 +142,15 @@ except ImportError:
             raise SyntaxError(e.message + ':\n' + template)
         result = namespace[typename]
 
-        # For pickling to work, the __module__ variable needs to be set to the frame
-        # where the named tuple is created.  Bypass this step in enviroments where
-        # sys._getframe is not defined (Jython for example) or sys._getframe is not
-        # defined for arguments greater than 0 (IronPython).
+        # For pickling to work, the __module__ variable needs to be set to the
+        # frame where the named tuple is created.  Bypass this step in
+        # enviroments where sys._getframe is not defined (Jython for example)
+        # or sys._getframe is not defined for arguments greater than 0
+        # (IronPython).
         try:
-            result.__module__ = _sys._getframe(1).f_globals.get('__name__', '__main__')
+            result.__module__ = _sys._getframe(1).f_globals.get('__name__',
+                                                                '__main__')
         except (AttributeError, ValueError):
             pass
 
         return result
-
-
-try:
-    import requests
-
-except ImportError:
-    class _response_container(object):
-        pass
-
-    def _perform_request(url, data=None, headers={}):
-        import urllib2
-        request = urllib2.Request(url, data=data, headers=headers)
-        response = urllib2.urlopen(request)
-        rv = _response_container()
-        rv.status_code = response.code
-        try:
-            rv.json = json.load(response)
-        except:
-            rv.json = None
-        return rv
-
-    class requests(object):
-        """ Implements a subset of the API provided by the `requests` module. """
-
-        @staticmethod
-        def get(url, params={}, headers={}):
-            if params:
-                import urllib
-                url += '?' + urllib.urlencode(params)
-            return _perform_request(url, None, headers)
-
-        @staticmethod
-        def post(url, data, headers={}):
-            import urllib
-            post_data = urllib.urlencode(data)
-            return _perform_request(url, post_data, headers)